fix: relocate settings out of flat entrypoints root so wxt build succeeds
