+++ wxt/entrypoints/utils/settings.ts
